working js, functional c++
C++ successfully parses the "event_logs.txt" file, but I could not implement inserting it into the logs.db file in time. 
Javascript implementation for displaying a database on a webpage is functional for the most part, successfully loading and displaying the information from the logs.db file that I inserted manually.